backend/requirements.txt
# Web Framework
flask==2.3.2
flask-cors==4.0.0
flask-restful==0.3.10

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
emoji==2.9.0

# ML/DL
torch>=2.0.0
transformers==4.36.2

# Data Collection
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3

# Others
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1